+++ proma/src/main/java/ai/platform/proma/service/AcountsService.java
@@ -3,8 +3,7 @@ package ai.platform.proma.service;
 import ai.platform.proma.domain.User;
 import ai.platform.proma.exception.ApiException;
 import ai.platform.proma.exception.ErrorDefine;
-import ai.platform.proma.repositroy.UserRepository;
-import lombok.AllArgsConstructor;
+import ai.platform.proma.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
