+++ proma/src/main/java/ai/platform/proma/service/OpenApiService.java
@@ -4,8 +4,7 @@ import ai.platform.proma.domain.Prompt;
 import ai.platform.proma.domain.User;
 import ai.platform.proma.exception.ApiException;
 import ai.platform.proma.exception.ErrorDefine;
-import ai.platform.proma.repositroy.PromptRepository;
-import ai.platform.proma.repositroy.UserRepository;
+import ai.platform.proma.repository.PromptRepository;
 import ai.platform.proma.security.openapi.OpenApiToken;
 import ai.platform.proma.security.openapi.OpenApiTokenProvider;
 import lombok.RequiredArgsConstructor;
