+++ proma/src/main/java/ai/platform/proma/service/PostService.java
@@ -7,8 +7,7 @@ import ai.platform.proma.dto.request.PostRequestDto;
 import ai.platform.proma.dto.response.*;
 import ai.platform.proma.exception.ApiException;
 import ai.platform.proma.exception.ErrorDefine;
-import ai.platform.proma.repositroy.*;
-import ai.platform.proma.security.LoginUser;
+import ai.platform.proma.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
